requirements.txt
selenium
python-dotenv
pillow
imagehash
scikit-image
numpy
opencv-python
openai
argparse